chore(setup)setup express application and static pages[finishes
