+++ .gitignore
@@ -1,1 +1,1 @@
-server/node_modules
+node_modules/
