+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "server",
+  "name": "auto-mart",
   "version": "1.0.0",
   "description": "An online market for automobile",
   "main": "index.js",
@@ -13,7 +13,10 @@
   },
   "keywords": [
     "auto",
-    "cars"
+    "cars",
+    "sales",
+    "model",
+    "vin"
   ],
   "author": "Opara Kelechi",
   "license": "ISC",
@@ -22,6 +25,7 @@
   },
   "homepage": "https://github.com/riskymind/Auto-Mart#readme",
   "dependencies": {
+    "body-parser": "^1.19.0",
     "express": "^4.17.1"
   },
   "devDependencies": {
@@ -29,8 +33,8 @@
     "@babel/core": "^7.4.5",
     "@babel/node": "^7.4.5",
     "@babel/preset-env": "^7.4.5",
+    "@babel/register": "^7.4.4",
     "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "^13.1.0",
     "nodemon": "^1.19.1"
   }
 }
